use BFP16_Convert_and_DistInnerProduct_bigEndian for distance calculation this function first generates a bf16 vectors and then calculates the IP distance into float.
add bf16 dist func to all hardwares optimization types. (in this commit it wasn't used yet)

the distance function  The function get float32 vectors. It first zero the LS bytes in each float in the vectors and then calcualte the distance as if they were regualar float vectors.
currently implemented only for IP, little endian

implement bm for bf16
in BM_VecSimIndex: if its bf16: initialize float32 flat index (for GT calculations), bf16 FLAT anf bf16 HNSW. don't insert vectors to the indices yet.

introduce BM_VecSimBF16dIndex which takes the raw vectors from a file and inserts them to all the indices. (we don't use serialized hnsw becuase the graph construction is also affected by the distance fucntion).

BM_VecSimCommon implemnts TopK_BF_bf16 that compute the recall of bf16 flat against float32 flat.

we have the run file of bf16 bm which defines raw_vectors_file path and does't initialize the mock
sizeof VecSimType_FP32_TO_BF16 is currently sizeof(float) for the datasize of the index
fix factory for bf type

+++ src/VecSim/utils/vec_utils.cpp
@@ -188,7 +188,7 @@ size_t VecSimType_sizeof(VecSimType type) {
     case VecSimType_INT64:
         return sizeof(int64_t);
     case VecSimType_FP32_TO_BF16:
-        return sizeof(bf16);
+        return sizeof(float);
     }
     return 0;
 }
